Adds code for delaunay_triangulation.cpp. This program reads a triangulation from stdin in OFF format, computes the preferred directions delaunay triangulation for directions u = (1,0) and v = (1,1), and outputs the triangulation to stdout in OFF format.
The code functions uses two algorithms to accomplish its purpose.
First, all halfedges are tested for flippability: if an edge is
unflippable, it is placed in a set called 'true_optimals', which contains
edges that will retain their optimality no matter what adjacent edges
are flipped. All halfedges that are found to be flippable are added
to a queue called 'sus'. This concludes the first algorithm.

The sus queue is then iterated over until empty. Each iteration, the
halfedge at the front of the queue is tested for optimality using the
preferred directions delaunay predicate from the ra::geometry::kernel
class. If the predicate is true, the halfedge is removed from the queue
and inserted into a set called 'optimals'. If the predicate fails, the
edge is flipped and any adjacent edges that are not members of the
true_optimals set are removed from the optimals set (if a part of it),
and placed back in the sus queue.

When the iterations are finished, the program outputs the preferred
directions delaunay triangulation to stdout.